refactor: replace async fun in round page
