+++ src/assets/style/style.scss
@@ -47,5 +47,4 @@ section {
 @import './common/common.scss';
 @import '../../view/pages/round-page/round-page.scss';
 @import '../../view/components/questionComponent/question.scss';
-@import '../../view/components/answersComponent/answers.scss';
 @import '../../view/components/nextRoudComponent/next-round.scss';
