+++ src/model/start-game.model.ts
@@ -1,7 +1,6 @@
 import { gameState } from "./state-game.model";
 import { wrapperFunction } from '../service/function-wrapper';
 import { appRender } from '../view/pageRender';
-
 export class StartGame {
 
     constructor() {
@@ -11,7 +10,6 @@ export class StartGame {
         gameState.createNewState();
         const data = gameState.appData;
         wrapperFunction(data).then((resolve) => {
-            console.log('Игра окончена');
             appRender.main.innerHTML = '';
             appRender.startPageRender();
         });
