--- src/service/custom-event-data-onchange.ts
@@ -1,10 +1,0 @@
-import { IAppData } from "../data/app-data.intarface";
-
-export class CustomEventOnChange {
-
-    constructor() { }
-
-    onChange: (data: IAppData) => CustomEvent = () => {
-        return new CustomEvent('onchange');
-    }
-}
