--- src/view/components/answersComponent/answers.scss
@@ -1,15 +1,0 @@
-.answers-list {
-    padding: 20px;
-
-    li {
-        padding: 5px;
-
-        input {
-            cursor: pointer;
-        }
-
-        label {
-            padding-left: 10px;
-        }
-    }
-}
