+++ src/view/pages/round-page/round-page.scss
@@ -16,6 +16,20 @@
             }
         }
     }
+}
 
+.answers-list {
+    padding: 20px;
 
+    li {
+        padding: 5px;
+
+        input {
+            cursor: pointer;
+        }
+
+        label {
+            padding-left: 10px;
+        }
+    }
 }
